Update HTTP client and server info
The story behind this change is the following.

Initially the 'http' module provides a client and a server both. In 2016
the alternative libcurl based client was implemented as the 'curl'
module.

In 2017 we [re-implemented][1] the libcurl based HTTP client as built-in
tarantool module (it was mainly to implement feedback daemon on top of
it).

Later in this year we [removed][2] the HTTP client from the 'http'
module. Now it contains only the server.

In 2019 we forked the curl repository into tarantool and moved the old
'curl' module into the 'tarantool-curl' repository (it was part of
[this][3] task).

Now the HTTP client is built-in and may be called with
`require('http.client')`. The server is in the external 'http' module
and available as `require('http.server')` just as before. This state is
described in the awesome-tarantool list now.

[1]: https://github.com/tarantool/tarantool/issues/2083
[2]: https://github.com/tarantool/http/issues/47
[3]: https://github.com/tarantool/tarantool/issues/4318